agent_framework/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.104
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nodejs==0.1.1
npm==0.1.1
numpy==1.26.4
ollama==0.3.1
openai==1.38.0
optional-django==0.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdfminer==20191125
pdfminer.six==20240706
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
